backend/api/webapp/requirements.txt
Flask>=1.1.2
gunicorn>=20.0.4
Jinja2>=2.11.2
joblib>=0.15.1
matplotlib>=3.2.2
pandas>=1.4.1
scikit-learn>=1.0.2
Werkzeug>=1.0.1
numpy>=1.21.5
exmol>=1.0.0
rdkit-pypi>=2021.9.5.1